D_01.03.24

+++ src/main/java/com/bogoliubova/training_service/dto/ServicesDto.java
@@ -3,10 +3,12 @@ package com.bogoliubova.training_service.dto;
 import lombok.Data;
 import lombok.Value;
 
+import java.math.BigDecimal;
+
 @Value
 //@Data
 public class ServicesDto {
     String type;
 
-    Double servicePrice;
+    BigDecimal servicePrice;
 }
